added some modifications

+++ utils/weather.py
@@ -5,9 +5,9 @@ class Weather(BaseModel):
     status : str
     description : str
     icon : str
-    temp : float
-    temp_min : float
-    temp_max : float
+    temp : int
+    temp_min : int
+    temp_max : int
     wind_speed : float
     dt : int
     sunrise : int
